Rename test files to correct names
